requirements.txt
torch
torchvision
torchmetrics
rasterio
numpy
pandas
tqdm
Pillow
h5py
